+++ tests/test_cli.py
@@ -21,6 +21,9 @@ raster_invalid_cog = os.path.join(
     os.path.dirname(__file__), "fixtures", "validate", "image_dec.tif"
 )
 raster_path_gcps = os.path.join(os.path.dirname(__file__), "fixtures", "slc.tif")
+raster_band_tags = os.path.join(
+    os.path.dirname(__file__), "fixtures", "cog_band_tags.tif"
+)
 
 
 @pytest.fixture(autouse=True)
@@ -526,5 +529,13 @@ def test_cogeo_info(runner):
         assert result.exit_code == 0
 
         result = runner.invoke(cogeo, ["info", raster_path_gcps])
+        assert not result.exception
+        assert result.exit_code == 0
+
+        result = runner.invoke(cogeo, ["info", raster_band_tags, "--json"])
+        assert not result.exception
+        assert result.exit_code == 0
+
+        result = runner.invoke(cogeo, ["info", raster_band_tags])
         assert not result.exception
         assert result.exit_code == 0